backend/pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^3.0"
python = "^3.8"
pyAesCrypt = "^6.0.0"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"
[tool.poetry.dev-dependencies]
